Move formatter => cli/formatter. Needed to move formatter.MemBytes => utils.MemBytes (helper class on memory display)
