+++ cli/cmd/compose/ps.go
@@ -27,7 +27,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/docker/compose-cli/api/client"
-	"github.com/docker/compose-cli/formatter"
+	"github.com/docker/compose-cli/cli/formatter"
 )
 
 func psCommand() *cobra.Command {
