+++ cli/formatter/logs.go
